requirements.txt
pytest-pudb
jedi
astor
GitPython

